ai-service/requirements.txt
fastapi
uvicorn
agno
google-genai
pydantic
python-dotenv
duckduckgo-search==6.4.2
